requirements.txt
streamlit
upstash-redis
langchain
langchain-community
langchain-together
faiss-cpu
huggingface-hub
python-docx
PyPDF2
pillow
pytesseract
numpy
tensorflow
keras
